pyproject.toml
[project]
name = "clif-mimic"
version = "0.3.1"
description = "Converting the MIMIC-IV dataset into the Common Longitudinal ICU data Format (CLIF)"
readme = "README.md"
requires-python = ">=3.10.5"
dependencies = [
    "duckdb>=1.3.0",
    "fastparquet>=2024.11.0",
    "pandas>=2.3.3",
    "pandera>=0.26.1",
    "pyarrow>=21.0.0",
    "sf-hamilton>=1.88.0",
]

[dependency-groups]
dev = [
    "auto-mix-prep>=0.2.0",
    "google-api-python-client>=2.184.0",
    "ipykernel>=6.30.1",
    "marimo>=0.16.4",
    "matplotlib>=3.10.7",
    "pytest>=8.4.2",
    "sqlglot>=27.18.0",
    "watchdog>=6.0.0",
]
